+++ src/main/java/by/forward/forward_system/core/jpa/repository/projections/ReviewProjectionDto.java
@@ -3,10 +3,12 @@ package by.forward.forward_system.core.jpa.repository.projections;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class ReviewProjectionDto {
-    private Integer orderTechNumber;
+    private BigDecimal orderTechNumber;
     private Long orderId;
     private Long reviewId;
     private String orderName;
